Separate buttons to sets user's goal
